Add files via upload
Q1)WAP for taking the string "yes " from the user till the user wants to find the area of a circle. here, the radius you have to take as input from the user.

Q2)WAP for finding given no is palindrome or not.

Q3)WAP for creating a mini calculator for performing a particular operation based on user choice. ex. if the user will select the addition operation then the output should be the addition of the given two nos.

Q4)WAP for taking a name from the user and displaying a welcome msg with the user name.